More matrix simplifications. No, really! All these format conversions I'm adding make things simpler!
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@1391 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/ParticleEntity.h
@@ -45,7 +45,7 @@ MA 02110-1301, USA.
 	OOColor			*color;
 	GLfloat			color_fv[4];
 	
-	double			alpha;
+	GLfloat			alpha;
 	double			time_counter;
 	
 	double			duration;
